Parameterize libvirt_cpu_mode
The libvirt_cpu_mode setting in nova.conf currently isn't exposed
via the compute::libvirt class, but can be very useful in nested
virtualization situations (as are common in development labs).
This patch adds a parameter and defaults it to 'false'.  If a
non-false parameter is provided, the setting will be added to the
[DEFAULT] section of the nova.conf file.

Change-Id: I0136f15621f616787aba1bbcbf68208bc8f9a1b1